requirements.txt
asgiref==3.6.0
attrs==22.2.0
cffi==1.15.1
colorama==0.4.6
contextlib2==21.6.0
coverage==7.2.1
cryptography==39.0.1
Django==4.1.7
django-cors-headers==3.13.0
django-extensions==3.2.1
django-ninja==0.21.0
django-ninja-extra==0.18.6
django-ninja-jwt==5.2.2
dnspython==2.3.0
email-validator==1.3.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
injector==0.20.1
ninja-schema==0.13.0
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
whitenoise==6.3.0
django-cleanup==7.0.0
